requirements.txt
asgiref==3.8.1
Django==5.2.1
djangorestframework==3.16.0
mysqlclient==2.2.7
PyMySQL==1.1.1
python-decouple==3.8
python-dotenv==1.1.0
sqlparse==0.5.3
tzdata==2025.2
